archive/requirements.txt
aenum==3.1.11
anyio==3.6.2
argcomplete==2.0.0
attrs==22.2.0
black==23.1a1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.0.1
click==8.1.3
datamodel-code-generator==0.16.0
dnspython==2.3.0
email-validator==1.3.0
filip==0.2.4
genson==1.2.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
igraph==0.9.8
importlib-resources==5.10.2
inflect==5.6.2
isodate==0.6.1
isort==5.11.4
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-spec==0.1.2
kafka-python==2.0.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.24.1
openapi-schema-validator==0.3.4
openapi-spec-validator==0.5.1
packaging==21.3
paho-mqtt==1.6.1
pandas==1.5.2
pandas-datapackage-reader==0.18.0
pathable==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
prance==0.22.11.4.0
pydantic==1.10.4
pyparsing==3.0.9
pyrsistent==0.19.3
PySnooper==1.1.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
PyYAML==6.0
rapidfuzz==2.13.7
rdflib==6.0.2
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
semver==2.13.0
six==1.16.0
sniffio==1.3.0
stringcase==1.2.0
texttable==1.6.7
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.4.0
urllib3==1.26.14
wget==3.2
